Various fixes + changes as suggested on PR by Luke
I moved some functions around so that they were in a more fitting
locatino, updated some comments, and coded a bit more defensively.
